+++ packages/apps-engine/src/server/bridges/IInternalUserBridge.ts
@@ -1,0 +1,8 @@
+import type { IUser, IUserCreationOptions } from '../../definition/users';
+
+export interface IInternalUserBridge {
+    create(data: Partial<IUser>, appId: string, options?: IUserCreationOptions): Promise<string>;
+    getAppUser(appId?: string): Promise<IUser | undefined>;
+    remove(user: IUser, appId: string): Promise<boolean>;
+    getActiveUserCount(): Promise<number>;
+}